# common/swagger-ui/swagger.yaml
swagger: '2.0'
info:
  description: This is an Node-service
  version: 1.0.0
  title: Swagger Node-service
  contact:
    email: Jerry.x.wang@newegg.com
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: localhost:3000
basePath: /v1
tags:
  - name: Customer
    description: Customer operation
schemes:
  - https
  - http
paths:
  /api/users/register:
    post:
      tags:
        - Customer
      summary: Register new user
      description: Register new user
      operationId: CustomerRegister
      consumes:
      - "application/json"
      - "application/xml"
      produces:
        - application/xml
        - application/json      
      parameters: 
        - in: body
          name: body
          description: Create customer operation
          required: true
          schema:
            $ref: '#/definitions/RegisterRequest'
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: '#/definitions/ServiceResponse'
  /api/users/login:
    post:
      tags:
        - Customer
      summary: Login a customer
      description: Login user
      operationId: CustomerLogin
      consumes:
      - "application/json"
      - "application/xml"
      produces:
        - application/xml
        - application/json
      parameters: 
        - in: body
          name: body
          description: Login customer operation
          required: true
          schema:
            $ref: '#/definitions/LoginRequest'
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/LoginResponse"
  /api/users/{loginName}:
    get:
      tags:
      - "Customer"
      summary: "Get customer login info"
      description: "Get customer info by login name"
      operationId: "getUserByName"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: "loginName"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: string
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Customer"
securityDefinitions:
  api_key:
    type: apiKey
    name: api_key
    in: header
definitions:
  Customer:
    type: object
    properties:
      loginName:
        type: string
      name:
        type: string
      phone:
        type: string
      email:
        type: string
      password:
        type: string
      avatar:
        type: string
      editUser:
        type: string
      countryCode:
        type: string
      companyCode:
        type: integer
        default: 1003
      languageCode:
        type: string
      inDate:
        type: string
      editDate:
        type: string
  RegisterRequest:
    type: object
    properties:
      loginName:
        type: string
      name:
        type: string
      phone:
        type: string
      email:
        type: string
      password:
        type: string
      avatar:
        type: string
      editUser:
        type: string
  LoginRequest: 
    type: object
    properties:
      loginName:
        type: string
      password:
        type: string
  LoginResponse: 
    type: object
    properties:
      message:
        type: string
      status:
        type: string
      token:
        type: string
  ServiceResponse:
    type: object
    properties:
      code:
        type: integer
        format: int32
      status:
        type: boolean
      message:
        type: string